+++ node/src/types/kernel/api/block/getBlockInfo/response.d.ts
@@ -1,0 +1,62 @@
+/**
+ * Copyright (C) 2023 SiYuan Community
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/**
+ * Gets the document information where the block in
+ */
+export interface IResponse {
+    /**
+     * status code
+     */
+    code: number;
+    data: IData;
+    /**
+     * status message
+     */
+    msg: string;
+}
+
+/**
+ * Response information
+ */
+export interface IData {
+    /**
+     * Notebook ID
+     */
+    box: string;
+    /**
+     * Document path, which needs to start with / and separate levels with /
+     * path here corresponds to the database path field
+     */
+    path: string;
+    /**
+     * Block ID without parent block
+     */
+    rootChildID: string;
+    /**
+     * Documenticon icon
+     */
+    rootIcon: string;
+    /**
+     * Document block ID
+     */
+    rootID: string;
+    /**
+     * Documenticon title
+     */
+    rootTitle: string;
+}